test: update style unit tests
